+++ extension/webview-ui/src/a11y/index.ts
@@ -1,0 +1,3 @@
+export { FOCUSABLE_SELECTOR, getFocusableElements, installFocusTrap } from "./focusTrap";
+export { LiveAnnouncer, PanelMain } from "./liveAnnouncer";
+export { usePrefersReducedMotion } from "./usePrefersReducedMotion";